requirements.txt
streamlit
openai
python-dotenv
sentence-transformers
faiss-cpu